db modify

+++ db/migrate/20170319125026_create_order_items.rb
@@ -2,9 +2,9 @@ class CreateOrderItems < ActiveRecord::Migration[5.1]
   def change
     create_table :order_items do |t|
       t.references :order, foreign_key: true
-      t.decimal :unit_price, precision: 12, scale: 3
+      t.integer :unit_price, precision: 12, scale: 3
       t.integer :quantity
-      t.decimal :total_price, precision: 12, scale: 3
+      t.integer :total_price, precision: 12, scale: 3
 
       t.timestamps
     end
