+++ db/migrate/20170319125916_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[5.1]
   def change
     create_table :products do |t|
       t.string :name
-      t.decimal :price, precision: 12, scale: 3
+      t.integer :price, precision: 12, scale: 3
       t.boolean :active
 
       t.timestamps
